[Fix/#93] SpotId Int -> Long

+++ feature/spot/src/main/java/com/acon/feature/spot/screen/spotlist/composable/SpotListScreen.kt
@@ -61,7 +61,7 @@ internal fun SpotListScreen(
     onResetFilter: () -> Unit = {},
     onCompleteFilter: (ConditionState, () -> Unit) -> Unit = { _, _ -> },
     onFilterBottomSheetShowStateChange: (Boolean) -> Unit = {},
-    onSpotItemClick: (id: Int) -> Unit = {},
+    onSpotItemClick: (id: Long) -> Unit = {},
 ) {
     val scrollState = rememberScrollState()
     val coroutineScope = rememberCoroutineScope()
